// src/app/types/types.ts
export type PatientFormData = {
    name: string;
    age: number;
    gender: string;
    address: string;
    firstConsultation: string;
    doctorName: string;
    symptoms: string;
    symptomDuration: string;
    severity: number;
    triggers: string;
    relievingFactors: string;
    chronicConditions: string[];
    pastSurgeries: string;
    familyHistory: string;
    onMedication: string;
    medicationList: string;
    supplements: string;
    allergies: string;
    sleepPattern: string;
    appetite: string;
    waterIntake: string;
    bowelMovements: string;
    physicalActivity: string[];
    height: number;
    weight: number;
    bp: string;
    pulse: string;
    temperature: string;
    spo2: string;
    diagnosis: string;
    investigations: string[];
    otherTests: string;
    reports: File[];
};

export type Consultation = {
  date: string;
  doctor: string;
  symptoms: string;
  duration: string;
  severity: number;
  triggers: string;
  relievingFactors: string;
  chronicConditions: string[];
  surgeries: string;
  familyHistory: string;
  takingMedication: string;
  medicationsList: string;
  supplements: string;
  allergies: string;
  sleep: string;
  appetite: string;
  waterIntake: string;
  bowel: string;
  activity: string;
  height: string;
  weight: string;
  bp: string;
  pulse: string;
  temperature: string;
  spo2: string;
  diagnosis: string;
  investigations: string[];
  otherTests: string;
};

export type Patient = {
  id: string;
  name: string;
  age: number;
  gender: string;
  address: string;
  consultations: Consultation[];
};